[Http] Move tests to the namespace that cooresponds with the main component structure.
